The whole point of this new constructor and prototype setup is so that we could have multiple instances of our Helper object. The prototype is just the key to take advantage of it.
To prove it all works, add var helper2 = new Helper() and pass it a different {dollar_sign}wrapper, like the footer on our page.

Since the footer doesn't have any rows that have weight on it, this should return zero. Log that: this.helper.calculateTotalWeight() and helper2.calculateTotalWeight().

Try that! Cool! 157.5 and of course, zero.