This is a Symfony form, but all this fanciness ultimately renders a good, old-fashioned form tag. Give the form another class: js-new-rep-log-form'.
Copy that and head into RepLogApp so we can attach a new listener. But wait... there is one problem: the  is actually the <table> element:

 77 lines  app/Resources/views/lift/index.html.twig
... lines 1 - 2
{% block body %}
  <div class='row'>
    <div class='col-md-7'>
... lines 6 - 12
      <table class='table table-striped js-rep-log-table'>
... lines 14 - 50
      </table>
      {{ include('lift/_form.html.twig') }}
    </div>
... lines 55 - 61
  </div>
{% endblock %}
... lines 64 - 77
And the form does not live inside of the <table>!

When you create little JavaScript applications like RepLogApp, you want the wrapper to be an element that goes around everything you need to manipulate.

Ok, no problem: let's move the js-rep-log-table class from the table itself to the div that surrounds everything.
Down below, I don't need to change anything here, but let's rename  to  for clarity.