Back in RepLogApp, add a success key to the AJAX call with a data argument: that will be the HTML we want to put on the page.
We need to replace all of this form code. I'll surround the form with a new element and give it a js-new-rep-log-form-wrapper class.

Back in success, use .closest() to find that, then replace its HTML with data.

We could have also used the replaceWith() jQuery function instead of targeting a parent element.
Sweet! Let's enjoy our work! Refresh and submit! Nice! But if I put 5 into the box and hit enter to submit a second time... it doesn't workBut down below, we can plainly see: calculateTotalWeight is a function! What the heck is going on? We'll fix that in a minute.