What about when we don't fail validation? In that case, we'll want to dynamically add a new row to the table. In other words, the AJAX call should once again return an HTML fragment: this time for a single <tr> row: this row right here.
To do that, we need to isolate it into its own template. Copy it, delete it, and create a new template: _repRow.html.twig. Paste the contents of this row in it.

Back in the main template, include this: lift/_repRow.html.twig.

Now that we've done this, we can render it directly in LiftController. We know that the form was submitted successfully if the code inside the {dollar_sign}form->isValid() block is executed. Instead of redirecting to another page, if this is AJAX, then return {dollar_sign}this->render('lift/_repRow.html.twig') and pass it the one variable it needs: repLog set to repLog.
And just by doing that, when we submit successfully, our AJAX endpoint returns the new <tr>.

But, our JavaScript code is already confused! It thought the new <tr> code was the error response, and replaced the form with it. Lame! Our JavaScript code needs to be able to distinguish between a successful request and one that failed with validation errors.

There's a perfectly standard way of doing this... and I was being lazy until now! On error, we should not return a 200 status code, and that's what the render() function gives us by default. When you return a 200 status code, it activates jQuery's success handler.