What about when we don't fail validation? In that case, we'll want to dynamically add a new row to the table. In other words, the AJAX call we should return a 400 status code, or really, anything that starts with a 4. To do that, add {dollar_sign}html = and then change render() to renderView().
This new method simply gives us the string HTML for the page. Next, return a new Response manually and pass it the content -  - and status code - 400.

As soon as we do that, the success function will not be called on errors. Instead, the error function will be called. For an error callback, the first argument is not the data from the response, it's a jqXHR object.

That's fine, because the response content is stored on jqXHR.responseText.

Now we can use the success function to add the new tr to the table. Before the AJAX call - to avoid any problems with the this variable - add {dollar_sign}tbody = this..find('tbody').

And in success, add .append(data).

That should do it!

Try it! Refresh the page! If we submit with errors, we get the errors! If we submit with something correct, a new row is added to the table. The only problem is that it doesn't update the total dynamically - that still requires a refresh.