But there's a much, much, much better way. AND, it comes with a fancy name: a delegate selector. Here's the idea, instead of attaching the listener to DOM elements that might be dynamically added to the page later, attach the listener to an element that will always be on the page. In our case, we know that this.{dollar_sign}wrapper will always be on the page.
Here's how it looks: instead of saying this.{dollar_sign}wrapper.find(), use this.{dollar_sign}wrapper.on() to attach the listener to the wrapper. Then, add an extra second argument, which is the selector for the element that you truly want to react to.

That's it! This works exactly the same as before. It just says:

Whenever a click event bubbles up to , please check to see if any elements inside of it with a js-delete-rep-log were also clicked. If they were, fire this function! And have a great day!
You know what else! When it calls handleRepLogDelete, the e.currentTarget is still the same as before: it will be the js-delete-rep-log link element. So all our code still works!