Let's update our AJAX call to go to this endpoint. In RepLogApp, down in handleNewFormSubmit, we need to somehow get that URL:
 102 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, $) {
... lines 4 - 24
  {dollar_sign}.extend(window.RepLogApp.prototype, {
... lines 26 - 61
    handleNewFormSubmit: function(e) {
... lines 63 - 67
     {dollar_sign}.ajax({
       url: .attr('action'),
... lines 70 - 79
     });
        }
  });
... lines 83 - 100
})(window, jQuery);

No problem! Find the form and add a fancy new data-url attribute set to path(), then the name of that route: rep_log_new.

Bam! Now, back in RepLogApp, before we use that, let's clear out all the code that actually updates our DOM: all the stuff related to updating the form with the form errors or adding the new row. That's all a todo for later.

But, do add a console.log('success') and console.log('error') so we can see if this stuff is working!