So how can we use this data to make actual HTML changes to the form? There are generally two different approaches. First, the simple way: parse the data by hand and manually use jQuery to add the necessary elements and classes. This is quick to do, but doesn't scale when things get really complex. The second way is to use a client-side template. We'll do the simple way first, but then use a client-side template for a more complex problem later.
And actually, there's a third way: which is to use a full front-end framework like ReactJS. We'll save that for a future tutorial.

Creating a Selectors Map

In _mapErrorsToForm, let's look at the error data and use it to add an error span below that field. Obviously, we need to use jQuery to find our .js-new-rep-log-form form element.

But wait! Way up in our constructor, we're already referencing this selector:
109 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}) {
  window.RepLogApp = function ({dollar_sign}wrapper) {
... lines 5 - 17
    this.{dollar_sign}wrapper.on(
... line 19
      '.js-new-rep-log-form',
... line 21
  );
};
... lines 25 - 107
})(window, jQuery);

It's no big deal, but I would like to not duplicate that class name in multiple places. Instead, add an _selectors property to our object. Give it a newRepForm key that's set to its selector.
Now, reference that with this._selectors.newRepForm.
Below in our function, do the same: var {dollar_sign}form = this.{dollar_sign}wrapper.find(this._selectors.newRepForm).