I left the span blank because it's cleaner to add the text on the next line: {dollar_sign}error.html(errorsData[fieldName]).
This jQuery object is now done! But it's not on the page yet. Add it with {dollar_sign}wrapper.append({dollar_sign}error). Also call {dollar_sign}wrapper.addClass('has-error').
Yes! Let's try it! Refresh and submit! There it is!

The only problem is that, once I finally fill in the field, the error message stays! AND, I get a second error message! Man, we gotta get this thing cleaned up!