The only problem was that, once I finally filled in the field, the error message stayed! AND, I got a second error message! Man, we gotta get this thing cleaned up!
No problem: at the top, use {dollar_sign}form.find() to find all the .js-field-error elements. And, remove those. Next, find all the form-group elements and remove the has-error class.

Refresh now, and re-submit! Errors! Fill in one... beautiful!

And if we fill in both fields, the AJAX call is successful, but nothing updates. Time to tackle that.