Update things to use the Underscore.js templating format. So, <%= totalWeightLifted %>. This is the print syntax, and I'm using a totalWeightLifted variable because eventually we're going to pass these keys to the template as variables: totalWeightLifted, reps, id, itemLabel and links.
Do the same thing to print out itemLabel. Keep going: the next line will be reps. And then use totalWeightLifted again... but make sure you use the right syntax!
But what about this data-url? We can't use the Twig path function anymore. But we can use this links._self key! That's supposed to be the link to where we can GET info about this RepLog, but because our API is well-built, it's also the URL to use for a DELETE request.

Great! Print out <%= links._self %>