Gosh, that's a nice template. Let's go use it! Find our _addRow() function. First, find the template text: {dollar_sign}('#js-rep-log-row-template').html(). Done! Our script tag trick is an easy way to store a template, but we could have also loaded it via AJAX. Winning!
Next, create a template object: var tpl = _.template(tplText).

That doesn't render the template, it just prepares it. Oh, and like before, my editor doesn't know what _ is... so I'll switch back to base.html.twig, press option+enter or alt+enter, and download that library. Much happier!

To finally render the template, add var html = tpl(repLog), where repLog is an array of all of the variables that should be available in the template.