Finally, celebrate by adding the new markup to the table: this.{dollar_sign}wrapper.find('tbody') and then .append({dollar_sign}.parseHTML(html)). The {dollar_sign}.parseHTML() function turns raw HTML into a jQuery object.
And since we have a new row, we also need to update the total weight. Easy! this.updateTotalWeightLifted().

Deep breath. Let's give this a shot. Refresh the page. I think we should lift our coffee cup ten times to stay in shape.

Ok, refresh and try it gain! This time, let's lift our coffee cup 20 times! It's alive!

If you watch closely, it's even updating the total weight at the bottom.

I love it! Except for the massive duplication: it's a real bummer to have the row template in two places. Let me show you one way to fix this.