Back in index.html.twig, find the <tbody> and empty it entirely: we do not need to render this stuff on the server anymore. In fact, we can even delete our _repRow.html.twig template entirely!
Let's keep celebrating: inside of LiftController - which renders index.html.twig - we don't need to pass in the repLogs or totalWeight variables to Twig: these will be filled in via JavaScript. Delete the totalWeight variable from Twig

If you refresh the page now, we've got a totally empty table. Perfect.