Effectively $.ajax() returns an object that has all the functionality of a Promise plus a few additional methods. The only methods that a true Promise has on it are .then() and .catch(), for when a promise is rejected, or fails. But jQuery's object also has .always(), .fail(), .done() and others that you can see inside what they call their 'deferred object'.
The story here is that jQuery implemented this functionality before the Promise object was a standard. You could use any of these methods, but instead, I want to focus on treating what we get back from jQuery as a pure Promise object. I want to pretend that these other methods don't exist, and only rely on .then() and .catch().

In other words, I'm saying:

Don't rely on .done(), just use .then(), which is the method you would use with any other library that implements Promises.
Modifying the Value in .then

Ok, go back and refresh now. When we submit, both handlers are still called. But woh! Check this out: our first data prints out correctly... but the second one is undefined?

If you look back at the Promise documentation, this makes sense. It says:

.then() appends a fulfillment handler on the Promise and returns a new Promise resolving to the return value of the called handler.
Ah, so when we add the second .then(), that's not being attached to the original Promise, that's being attached to a new Promise that's returned from the first .then(). And according to the rules, the value for that new Promise is equal to whatever we return from the first.