If you look back at the Promise documentation, this makes sense. It says:
.then() appends a fulfillment handler on the Promise and returns a new Promise resolving to the return value of the called handler.
Ah, so when we add the second .then(), that's not being attached to the original Promise, that's being attached to a new Promise that's returned from the first .then(). And according to the rules, the value for that new Promise is equal to whatever we return from the first.

Ok, so let's prove that's the case: return data.

Back in the browser, it works! Both handlers are passed the same data.

But what about handling failures? Oh, that's pretty crazy.