Yay! Let's complicate things!
Our AJAX call works really well, because when we make an AJAX call to create a new RepLog, our server returns all the data for that new RepLog. That means that when we call .then() on the AJAX promise, we have all the data we need to call _addRow() and get that new row inserted!

 166 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}, Routing) {
... lines 4 - 26
  {dollar_sign}.extend(window.RepLogApp.prototype, {
... lines 28 - 77
    handleNewFormSubmit: function(e) {
... lines 79 - 86
      this._saveRepLog(formData)
        .then(function(data) {
          self._clearForm();
          self._addRow(data);
... lines 91 - 93
      });
    },
... lines 96 - 145
  });
... lines 147 - 164
})(window, jQuery, Routing);
Too easy: so let's make it harder!