Pretend that we don't have full control over our API. And instead of returning the RepLog data from the create endpoint - which is what this line does - it returns an empty response.
Passing null means no response content, and 204 is just a different status code used for empty responses - that part doesn't make any difference.

Now head over and fill out the form successfully. Whoa!

Yep, it blew up - that's not too surprising: we get an error that says:

totalWeightLifted is not defined.
And if you look closely, that's coming from underscore.js. This is almost definitely an error in our template. We pass the response data - which is now empty - into ._addRow():
166 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}, Routing) {
... lines 4 - 26
  {dollar_sign}.extend(window.RepLogApp.prototype, {
... lines 28 - 77
    handleNewFormSubmit: function(e) {
... lines 79 - 87
      .then(function(data) {
... line 89
        self._addRow(data);
... lines 91 - 93
      });
    },
... lines 96 - 145
  });
... lines 147 - 164
})(window, jQuery, Routing);

And that eventually becomes the variables for the template:

 166 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}, Routing) {
... lines 4 - 26
  {dollar_sign}.extend(window.RepLogApp.prototype, {
... lines 28 - 136
    _addRow: function(repLog) {
      var tplText = {dollar_sign}('#js-rep-log-row-template').html();
      var tpl = _.template(tplText);
      var html = tpl(repLog);
      this.{dollar_sign}wrapper.find('tbody').append({dollar_sign}.parseHTML(html));
      this.updateTotalWeightLifted();
    }
  });
... lines 147 - 164
})(window, jQuery, Routing);
An empty response means that no variables are being passed. Hence, totalWeightLifted is not defined.

But check this out: there's a second error:

JSON Exception: unexpected token.

A catch Catches Errors

This is coming from RepLogApp.js, line 94. Woh, it's coming from inside our .catch() handler:

 166 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}, Routing) {
... lines 4 - 26
  {dollar_sign}.extend(window.RepLogApp.prototype, {
... lines 28 - 77
    handleNewFormSubmit: function(e) {
... lines 79 - 86
      this._saveRepLog(formData)
        .then(function(data) {
... lines 89 - 90
         }).catch(function(jqXHR) {
           var errorData = JSON.parse(jqXHR.responseText);
           self._mapErrorsToForm(errorData.errors);
         });
      },
... lines 96 - 145
    });
... lines 147 - 164
})(window, jQuery, Routing);
Now, as we understand it, our catch should only be called when our Promise fails, in other words, when we have an AJAX error. But in this case, the server returns a 204 status code - that is a successful status code. So why is our catch being called?

Here's the deal: in reality, .catch() will be called if your Promise is rejected, or if a handler above it throws an error. Since our .then() calls _addRow() and that throws an exception, this ultimately triggers the .catch(). Again, this works a lot like the try-catch block in PHP!

 Tip
There are some subtle cases when throwing an exception inside asynchronous code won't trigger your .catch(). The Mozilla Promise Docs(https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/catch#Gotchas_when_throwing_errors) discuss this!