Let's make an AJAX call to that: copy the jqXHR line. Then, add our favorite {dollar_sign}.ajax() and set the URL to that header. Add a .then() to this Promise with a data argument.
Finally, this should be the RepLog data.

To check things, add console.log('now we are REALLY done') and also console.log(data) to make sure it looks right.

Ok, refresh and fill out the form. Ignore the errors, because there's our message and the correct data!

Ok, now we can just return this somehow, right? Wait, that's not going to work... When we return the main {dollar_sign}.ajax(), that Promise is resolved - meaning finished - the moment that the first AJAX call is made. You can see that because the errors from the handlers happen first, and then the second AJAX call finishes.

Somehow, we need to return a Promise that isn't resolved until that second AJAX call finishes.

There are two ways to do this - we'll do the harder way... because it's a lot more interesting - but I'll mention the other way at the end.