Could we use a Promise?
What we need to do is create our own Promise object, and take control of exactly when it's resolved and what value is passed back.

If you look at the Promise documentation, you'll find an example of how to do this: new Promise() with one argument: a function that has resolve and reject arguments. I know, it looks a little weird.

Inside of that function, you'll put your asynchronous code. And as soon as it's done, you'll call the resolve() function and pass it whatever value should be passed to the handlers. If something goes wrong, call the reject() function. This is effectively what jQuery is doing right now inside of its $.ajax() function.

There's one quick gotcha: not all browsers support the Promise object. But, no worries! Google for 'JavaScript Promise polyfill CDN'.

A polyfill is a library that gives you functionality that's normally only available in a newer version of your language, JavaScript in this case. PHP also has polyfills: small PHP libraries that backport newer PHP functionality.

This polyfill guarantees that the Promise object will exist in JavaScript. If it's already supported by the browser it uses that. But if not, it adds it.

Copy the es6-promise.auto.min.js path. In the next tutorial, we'll talk all about what that es6 part means. Next, go into app/Resources/views/base.html.twig and add a script tag with src='https://cdnjs.cloudflare.com/ajax/libs/es6-promise/4.0.5/es6-promise.auto.min.js'

Now our Promise object is guaranteed!