In _saveRepLog, create and return a new Promise, passing it the 1 argument it needs: a function with resolve and reject arguments.
Move all of our AJAX code inside.

Now, all we need to do is call resolve() when our asynchronous work is finally resolved. This happens after the second AJAX call. Great! Just call resolve() and pass it data.

Finally, the RepLog data should once again be passed to the success handlers!

Go back now and refresh. Watch the total at the bottom: lift the big fat cat 10 times and... boom! The new row was added and the total was updated. It worked!

This is huge! Our _saveRepLog function previously returned a jqXHR object, which implements the Promise interface. Now, we've changed that to a real Promise, and our code that calls this function didn't need to change at all. The .then() and .catch() work exactly like before. Ultimately, before and after this change, _saveRepLog() returns a promise whose value is the RepLog data.