Of course, we also need to call reject, which should happen if the original AJAX call has a validation error. If you fill out the form blank now, we can see the 400 error, but it doesn't call our .catch() handler.
No problem: after .then(), add a .catch() to handle the AJAX failure. Inside that, call reject() and pass it jqXHR: the value that our other .catch() expects.

We could also add a .catch() to the second AJAX call, but this should never fail under normal circumstances, so I think that's overkill.

Refresh again! And try the form blank. Perfect! But, we can get a little bit fancier.